+++ worker-node.tf
@@ -23,8 +23,7 @@ packages:
 users:
   - name: cluster
     ssh-authorized-keys:
-      - ${trimspace(var.ssh_public_key)}
-      - ${trimspace(tls_private_key.worker_node.public_key_openssh)}
+      ${indent(6, trimspace(yamlencode(local.worker_authorized_keys)))}
     sudo: ALL=(ALL) NOPASSWD:ALL
     shell: /bin/bash